refactor(back): SubscriptionStatus(TRIAL) 정책 반영 & 대시보드 월 지출 계산 로직 개선 - 월 지출 합계를 전체 구독 기준 → 이번 달 평가된 구독 기준으로 변경 - TRIAL 상태 구독은 비용 0으로 계산하도록 처리

+++ backend/src/main/java/com/back/domain/dashboard/dto/DashboardResponse.java
@@ -4,7 +4,7 @@ import com.back.domain.evaluation.enums.EvaluationStatus;
 import java.util.List;
 
 public record DashboardResponse(
-        int totalMonthlyCost,          // 이번 달 총 구독 지출액
+        long totalMonthlyCost,          // 이번 달 평가된 구독 총액
         int totalAnnualWasteEstimate,  // 이번 달 기준 연간 예상 낭비액 합계
         List<SubscriptionSummary> subscriptions
 ) {
